Split core into two modules
